Update on "Make torch.device usable as a context manager"
Fixes https://github.com/pytorch/pytorch/issues/82296
Fixes https://github.com/pytorch/pytorch/issues/27878
Fixes https://github.com/pytorch/pytorch/issues/260

Open to bikeshedding the module name.  Open to bikeshedding if any of
the identifiers should get reexporting from torch top level.  Open to
bikeshedding the global setter API name / location.


[ghstack-poisoned]
