+++ test/test_utils.py
@@ -23,7 +23,7 @@ from torch.testing._internal.common_methods_invocations import op_db
 import torch.cuda
 from torch.utils._pytree import tree_any, tree_all_only
 from torch.utils.checkpoint import checkpoint, checkpoint_sequential
-from torch.utils.device_mode import DeviceMode, set_default_tensor_device
+from torch import set_default_tensor_device
 import torch.utils.cpp_extension
 from torch.autograd._functions.utils import check_onnx_broadcast
 from torch.onnx.symbolic_opset9 import _prepare_onnx_paddings
@@ -804,14 +804,15 @@ class TestExtensionUtils(TestCase):
             torch._register_device_module('xpu', DummyXPUModule)
 
 
-class TestDeviceModeUtils(TestCase):
+class TestDeviceUtils(TestCase):
     def test_basic(self):
-        with DeviceMode('meta'):
+        with torch.device('meta') as dev:
             x = torch.empty(3, 3)
         self.assertEqual(x.device.type, 'meta')
+        self.assertEqual(dev, torch.device('meta'))
 
     def test_nn_module(self):
-        with DeviceMode('meta'):
+        with torch.device('meta'):
             m = nn.Linear(40, 50)
         self.assertEqual(m.weight.device.type, 'meta')
 
@@ -838,21 +839,21 @@ class TestDeviceModeUtils(TestCase):
                 (sample.input, sample.args, sample.kwargs)
             ):
                 continue
-            # Many OpInfos will explicitly pass in a device.  DeviceMode
+            # Many OpInfos will explicitly pass in a device.  DeviceContext
             # will respect device if it is explicitly specified.  To test
-            # DeviceMode, we have to remove the device kwarg in this case.
+            # DeviceContext, we have to remove the device kwarg in this case.
             # NB: Can't pass None to sample_inputs, the function can't
             # handle it.
             kwargs = sample.kwargs.copy()
             kwargs.pop('device', None)
-            with DeviceMode('meta'):
+            with torch.device('meta'):
                 r = func(sample.input, *sample.args, **kwargs)
             self.assertTrue(
                 tree_all_only(torch.Tensor, lambda x: x.device.type == 'meta', r)
             )
 
 
-instantiate_device_type_tests(TestDeviceModeUtils, globals())
+instantiate_device_type_tests(TestDeviceUtils, globals())
 
 
 class TestCppExtensionUtils(TestCase):
